refactor(api): rename certification-center-invitation-controller.js
to certification-center-invitation.controller.js
