+++ api/src/team/application/routes.js
@@ -2,7 +2,7 @@ import Joi from 'joi';
 
 import { securityPreHandlers } from '../../shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../shared/domain/types/identifiers-type.js';
-import { certificationCenterInvitationController } from './certification-center-invitation-controller.js';
+import { certificationCenterInvitationController } from './certification-center-invitation.controller.js';
 
 const register = async function (server) {
   server.route([
